+++ Models/IndividualTraining.cs
@@ -10,7 +10,7 @@ namespace WebApi.Models
 
         public MembershipReceipt MembershipReceipt { get; set; }
 
-        public int PayedHours { get; set; }
+        public double PayedHours { get; set; }
 
         public decimal Price { get; set; }
